fz.STS_PRS_251_presence and tz.STS_PRS_251_beep
